+++ src/Components/ButtonIconForInterface.jsx
@@ -1,0 +1,17 @@
+import { doI18n } from "pithekos-lib";
+
+function ButtonIconForInterface(i18nRef, Icon, onClickFunction, title) {
+  <Tooltip
+    title={doI18n(`pages:core-dashboard:${title}`, i18nRef.current)}
+    disableInteractive
+    placement="right"
+  >
+    <IconButton
+      onClick={() => {
+        onClickFunction();
+      }}
+    >
+      <Icon />
+    </IconButton>
+  </Tooltip>;
+}
